+++ CHANGELOG.md
@@ -14,7 +14,7 @@
   - Sheets not on an event now have their creation date listed on the subject
     page
 - **PATS Changes**
-  - Added Signal Quality Grades table to Data Quality export
+  - Added Signal Quality Grades table and graph to Data Quality export
 
 ### Refactoring
 - Removed deprecated time parsing code